requirements.txt
flask
functions-framework
google-api-core
google-cloud-translate
gunicorn
google-auth